+++ javacup-dashboard/src/main/java/io/cutalab/javacup/dashboard/views/MainLayout.java
@@ -65,6 +65,7 @@ public class MainLayout extends AppLayout {
         navigation.addItem(new SideNavItem("Processes", ProcessesView.class));
         navigation.addItem(new SideNavItem("Current JVM Metrics", CurrentJvmMetricsView.class));
         navigation.addItem(new SideNavItem("Metric Samples", MetricSamplesView.class));
+        navigation.addItem(new SideNavItem("Donate", DonationsView.class));
 
         Scroller scroller = new Scroller(navigation);
         addToDrawer(scroller);